pyproject.toml
[tool.poetry]
name = "nonebot-plugin-ncupdate"
version = "0.1.4"
description = "A nonebot plugin for update NapCat"
authors = ["tianyisama"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/tianyisama"
repository = "https://github.com/tianyisama/nonebot-plugin-ncupdate"
[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "<3.0.0,>=2.2.0"
nonebot-adapter-onebot = "^2.1.1"
psutil = ">=5.9.0,<6.0.0"
aiofiles = ">=23.2.1"
httpx = ">=0.23.3"
[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"